Rearrange code layout, add testmediainfo command, improve media tool parsing

+++ PlexCleaner/FfMpegToolJsonSchema.cs
@@ -1,6 +1,10 @@
+#region
+
 using System.Collections.Generic;
 using System.Text.Json;
 using System.Text.Json.Serialization;
+
+#endregion
 
 // Convert JSON file to C# using app.quicktype.io
 // Set language, framework, namespace, list
@@ -45,7 +49,7 @@ public class FfMpegToolJsonSchema
     public class Track
     {
         [JsonPropertyName("index")]
-        public int Index { get; set; }
+        public long Index { get; set; }
 
         [JsonPropertyName("codec_name")]
         public string CodecName { get; set; } = "";
